feat: text color and publish tailwind config
