+++ src/style.scss
@@ -1,4 +1,3 @@
-@import "styles/typography.scss";
 @import "styles/colors.scss";
 
 @tailwind base;
