--- src/styles/typography.scss
@@ -1,55 +1,0 @@
-.text-body-1 {
-    @apply text-base;
-}
-
-.text-body-2 {
-    @apply text-sm;
-}
-
-.text-subtitle-1 {
-    @apply text-base leading-7;
-}
-
-.text-subtitle-2 {
-    @apply text-sm leading-5 font-medium;
-}
-
-.text-subtitle-2 {
-    @apply text-sm leading-5 font-medium;
-}
-
-.text-caption {
-    @apply text-xs leading-5;
-}
-
-.text-overline {
-    @apply text-xs leading-8 uppercase;
-}
-
-.text-h1 {
-    @apply text-8xl font-light;
-    line-height: 7rem;
-}
-
-.text-h2 {
-    @apply text-6xl font-light;
-    line-height: 4.5rem;
-}
-
-.text-h3 {
-    @apply text-5xl;
-    line-height: 3.5rem;
-}
-
-.text-h4 {
-    @apply text-4xl font-semibold;
-    line-height: 2.625rem;
-}
-
-.text-h5 {
-    @apply text-2xl leading-8;
-}
-
-.text-h6 {
-    @apply text-xl leading-8;
-}
